clean up schema.sql
